레이아웃 수정 및 api 연동 시작
