+++ src/hooks/checkToken.tsx
@@ -1,0 +1,4 @@
+export const useCheckToken = () => {
+  const token = localStorage.getItem("access_token");
+  return { isLoggedIn: !!token };
+};
