+++ src/layout/layout.tsx
@@ -2,8 +2,21 @@ import { Outlet } from "react-router-dom";
 import Sidebar from "./sidebar/sidebar";
 import * as S from "./layoutStyled";
 import Top from "./top/top";
+import { useNavigate } from "react-router-dom";
+import { useCheckToken } from "../hooks/checkToken";
+import { useEffect } from "react";
 
 const Layout = () => {
+  const { isLoggedIn } = useCheckToken();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (!isLoggedIn) {
+      alert("로그인해 주세요");
+      navigate("/auth");
+    }
+  }, [isLoggedIn, navigate]);
+
   return (
     <S.Wrapper>
       <Top />
